Changed focus to UI
Since there was big problem with File uploading, and that part isn't
set to be built for a bit, I decided to turn my attention to UI.

That doesn't mean the UI is good yet.  I just have a rough version of
using the toolbar instead of adding a menu for the PhoneGap extension.

Also Fixed GH-1.